Extract logic to utils file
